:sparkles: Add OpCodeFactory as new file
