+++ BitcoinKit/Opcode.swift
@@ -188,7 +188,7 @@ public class OpCode: OpCodeProtocol {
 
     public static let OP_INVALIDOPCODE: OpCode = OpInvalidOpCode()
 
-    fileprivate static let list: [OpCode] = [
+    internal static let list: [OpCode] = [
         OP_0,
         OP_FALSE,
         OP_PUSHDATA1,
@@ -306,52 +306,6 @@ public class OpCode: OpCodeProtocol {
     ]
 
     fileprivate init() {}
-}
-
-public struct OpCodeFactory {
-    public static func get(with value: UInt8) -> OpCode {
-        guard let item = (OpCode.list.first { $0.value == value }) else {
-            return OpInvalidOpCode()
-        }
-        return item
-    }
-
-    public static func get(with name: String) -> OpCode {
-        guard let item = (OpCode.list.first { $0.name == name }) else {
-            return OpInvalidOpCode()
-        }
-        return item
-    }
-
-    // Returns OP_1NEGATE, OP_0 .. OP_16 for ints from -1 to 16.
-    // Returns OP_INVALIDOPCODE for other ints.
-    public static func opcodeForSmallInteger(smallInteger: Int) -> OpCode {
-        switch smallInteger {
-        case -1:
-            return OpCode.OP_1NEGATE
-        case 0:
-            return OpCode.OP_0
-        case 1...16:
-            return get(with: OpCode.OP_1.value + UInt8(smallInteger - 1))
-        default:
-            return OpCode.OP_INVALIDOPCODE
-        }
-    }
-
-    // Converts opcode OP_<N> or OP_1NEGATE to an integer value.
-    // If incorrect opcode is given, Int.max is returned.
-    public static func smallIntegerFromOpcode(opcode: OpCode) -> Int {
-        switch opcode {
-        case .OP_1NEGATE:
-            return -1
-        case .OP_0:
-            return 0
-        case (.OP_1)...(.OP_16):
-            return Int(opcode.value - OpCode.OP_1.value - 1)
-        default:
-            return Int.max
-        }
-    }
 }
 
 extension OpCode: Comparable {
